feat: fully adopt Credit value object — wire Credit.SCALE into ledger, schema, and all call sites (WOP-1059)
- Credit.SCALE = 1_000_000_000 (nanodollars): fromCents multiplies by 1e7, toCents divides by 1e7
- creditColumn Drizzle custom type stores Credit as bigint in PostgreSQL
- Migration 0049: drop credit columns and recreate as bigint
- ICreditLedger.balance() returns Promise<Credit>, credit()/debit() accept Credit amounts
- CreditTransaction: amount and balanceAfter are Credit instances (not raw ints)
- DividendCronResult.pool/perUser are Credit instances
- All route handlers serialize with .toCents() before sending JSON
- PGlite SUM(bigint) returns float — wrap with Math.round() before Credit.fromRaw()
- All test files updated to use Credit value objects throughout

+++ src/db/credit-column.ts
@@ -6,7 +6,7 @@ export const creditColumn = customType<{
   driverData: number;
 }>({
   dataType() {
-    return "integer";
+    return "bigint"; // nanodollar values exceed int4 range
   },
   toDriver(value: Credit): number {
     return value.toRaw();
